add Method Virify by Bi Motorists,change language string TypeInfringements and add price

+++ src/main/java/ao/multaplus/typeInfringement/dtos/TypeInfringementsUpdateDto.java
@@ -6,7 +6,7 @@ public record TypeInfringementsUpdateDto(
         Long id,
         String type,
         String description,
-        Long price,
+        float price,
         StateSenderDto state
 ) {
 }
